requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
docx==0.2.4
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
lxml==5.1.0
openai==1.12.0
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
PyPDF2==3.0.1
python-docx==1.1.0
pytz==2024.1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
